+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.4.3 - 2018-11-22
 
 * [FEATURE] - Improve caret position in columns
+* [FEATURE] - Update react-select to v2
 
 ## v1.4.2 - 2018-11-16